+++ semantic_navigation_msgs/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package semantic_navigation_msgs
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.0 (03-11-2023)
+------------------
+* Added SemanticRegions.srv action file.
+
 1.0.0 (25-07-2023)
 ------------------
 * Move from semantic_goals_generator (2.2.0)